+++ src/main/java/de/maxhenkel/voicechat/gui/widgets/MicActivationButton.java
@@ -3,6 +3,7 @@ package de.maxhenkel.voicechat.gui.widgets;
 import de.maxhenkel.voicechat.VoicechatClient;
 import de.maxhenkel.voicechat.voice.client.MicrophoneActivationType;
 import net.minecraft.client.gui.components.AbstractButton;
+import net.minecraft.client.gui.narration.NarrationElementOutput;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
 
@@ -34,5 +35,10 @@ public class MicActivationButton extends AbstractButton {
         VoicechatClient.CLIENT_CONFIG.microphoneActivationType.set(type);
         VoicechatClient.CLIENT_CONFIG.microphoneActivationType.save();
         updateText();
+    }
+
+    @Override
+    public void updateNarration(NarrationElementOutput narrationElementOutput) {
+        this.defaultButtonNarrationText(narrationElementOutput);
     }
 }