+++ src/main/java/de/maxhenkel/voicechat/gui/widgets/MicTestButton.java
@@ -8,6 +8,7 @@ import de.maxhenkel.voicechat.voice.client.DataLines;
 import de.maxhenkel.voicechat.voice.client.MicThread;
 import de.maxhenkel.voicechat.voice.common.Utils;
 import net.minecraft.client.gui.components.AbstractButton;
+import net.minecraft.client.gui.narration.NarrationElementOutput;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.network.chat.TranslatableComponent;
 
@@ -92,6 +93,11 @@ public class MicTestButton extends AbstractButton {
             return;
         }
         micThread.setMicrophoneLocked(locked);
+    }
+
+    @Override
+    public void updateNarration(NarrationElementOutput narrationElementOutput) {
+        this.defaultButtonNarrationText(narrationElementOutput);
     }
 
     private class VoiceThread extends Thread {