+++ changelog.md
@@ -1,1 +1,1 @@
-- Fixed volume adjusting screen
+- Updated to 1.17-pre1